moved bookstore related dtos to their own BookStore folder
